update: modify TableColumnsConfig

+++ optic/dialog/table_columns_config.py
@@ -2,10 +2,12 @@ from __future__ import annotations
 from ..type_definitions import *
 from PyQt5.QtWidgets import QDialog, QVBoxLayout, QHBoxLayout, QTableWidget, QTableWidgetItem, QComboBox, QLineEdit
 from PyQt5.QtCore import Qt
+from PyQt5.QtWidgets import QMainWindow, QMessageBox
 from ..manager.widget_manager import WidgetManager
 from ..manager.init_managers import initManagers
 from ..config.constants import TableColumnConfigDialog_Config
 from ..utils.table_utils import deleteSelectedRows, addRow
+import copy
 
 # Table Columns Config
 class TableColumnConfigDialog(QDialog):
@@ -17,6 +19,7 @@ class TableColumnConfigDialog(QDialog):
             ):
         super().__init__(parent)
         self.widget_manager = initManagers(WidgetManager())
+        self.table_columns_tmp = copy.copy(table_columns) # copy to prevent changing original table_columns
         self.table_columns = table_columns
 
         window_settings = gui_defaults.get("WINDOW_SETTINGS_TABLE_COLUMNS_CONFIG", {})
@@ -52,32 +55,52 @@ class TableColumnConfigDialog(QDialog):
 
         self.bindFuncAllWidget()
 
+    # get table_columns info from the selected row
+    def getTableColumnsInfo(self, row: int) -> Tuple[str, Dict[str, Any]]:
+        col_name = self.widget_manager.dict_table["table_columns"].cellWidget(row, 0).text()
+        col_info = self.table_columns.getColumns()[col_name]
+        return col_name, col_info
+
     def setupConfigTable(self):
         columns = self.table_columns.getColumns()
         self.widget_manager.dict_table["table_columns"].setRowCount(len(columns))
         
+        # i: row
         for i, (col_name, col_info) in enumerate(columns.items()):
             # Column Name
             self.widget_manager.makeWidgetLineEdit(key=f"col_name_{i}", text_set=col_name)
             self.widget_manager.dict_table["table_columns"].setCellWidget(i, 0, self.widget_manager.dict_lineedit[f"col_name_{i}"])
-            
+            if col_info.get("name_fixed", False): # uneditable
+                self.widget_manager.dict_lineedit[f"col_name_{i}"].setReadOnly(True)
+
             # Type
-            self.widget_manager.makeWidgetComboBox(key=f"type_{i}")
-            self.widget_manager.dict_combobox[f"type_{i}"].addItems(TableColumnConfigDialog_Config.COMBO_ITEMS)
-            self.widget_manager.dict_combobox[f"type_{i}"].setCurrentText(col_info[f'type'])
-            self.widget_manager.dict_table["table_columns"].setCellWidget(i, 1, self.widget_manager.dict_combobox[f"type_{i}"])
+            if col_info.get("name_fixed", False): # uneditable
+                self.widget_manager.makeWidgetLineEdit(key=f"type_{i}", text_set=col_info[f'type'])
+                self.widget_manager.dict_table["table_columns"].setCellWidget(i, 1, self.widget_manager.dict_lineedit[f"type_{i}"])
+                self.widget_manager.dict_lineedit[f"type_{i}"].setReadOnly(True)
+            else:
+                self.widget_manager.makeWidgetComboBox(key=f"type_{i}")
+                self.widget_manager.dict_combobox[f"type_{i}"].addItems(TableColumnConfigDialog_Config.COMBO_ITEMS)
+                self.widget_manager.dict_combobox[f"type_{i}"].setCurrentText(col_info[f'type'])
+                self.widget_manager.dict_table["table_columns"].setCellWidget(i, 1, self.widget_manager.dict_combobox[f"type_{i}"])
             
             # Width
             self.widget_manager.makeWidgetLineEdit(key=f"width_{i}", text_set=str(col_info['width']))
             self.widget_manager.dict_table["table_columns"].setCellWidget(i, 2, self.widget_manager.dict_lineedit[f"width_{i}"])
 
     def deleteSelectedTableColumns(self):
-        deleteSelectedRows(self.widget_manager.dict_table["table_columns"])
+        selected_row = sorted(set(index.row() for index in self.widget_manager.dict_table["table_columns"].selectedIndexes()), reverse=True)[0]
+        selected_col_name, selected_col_info = self.getTableColumnsInfo(selected_row)
+        if not selected_col_info.get("removable", False):
+            QMessageBox.warning(self, "Warning", f"{selected_col_name} is not removable !")
+        else:
+            deleteSelectedRows(self.widget_manager.dict_table["table_columns"])
+            self.updateTableColumnsTmp()
 
     def addNewTableColumn(self):
         row = addRow(self.widget_manager.dict_table["table_columns"])
         
-        # 新しい行にウィジェットを追加
+        # add new widgets
         self.widget_manager.makeWidgetLineEdit(key=f"col_name_{row}", text_set=TableColumnConfigDialog_Config.DEFAULT_PARAMS[0])
         self.widget_manager.dict_table["table_columns"].setCellWidget(row, 0, self.widget_manager.dict_lineedit[f"col_name_{row}"])
         
@@ -96,34 +119,55 @@ class TableColumnConfigDialog(QDialog):
 
         for row in range(table.rowCount()):
             col_name = table.cellWidget(row, 0).text()
-            col_type = table.cellWidget(row, 1).currentText()
+            try:
+                col_type = table.cellWidget(row, 1).currentText() # combo box
+            except:
+                col_type = table.cellWidget(row, 1).text() # line edit
             col_width = int(table.cellWidget(row, 2).text())
 
-            column_info = {
+            # original col_info
+            col_info = self.table_columns_tmp.getColumns().get(col_name, {})
+            # col_info for update
+            col_info_tmp = {
                 "order": row,
                 "type": col_type,
-                "width": col_width
+                "width": col_width,
             }
 
             if col_type == "celltype":
                 if not celltype_found:
-                    column_info["default"] = True
+                    col_info_tmp["default"] = True
                     celltype_found = True
                 else:
-                    column_info["default"] = False
+                    col_info_tmp["default"] = False
             elif col_type == "checkbox":
-                column_info["default"] = False
+                col_info_tmp["default"] = False
 
-            table_columns[col_name] = column_info
+            if col_info.get("name_fixed", False):
+                col_info_tmp["name_fixed"] = col_info["name_fixed"]
+            if col_info.get("removable", False):
+                col_info_tmp["removable"] = col_info["removable"]
+            if col_info.get("editable", False):
+                col_info_tmp["editable"] = col_info["editable"]
+            if col_info.get("default", False):
+                col_info_tmp["default"] = col_info["default"]
+
+            table_columns[col_name] = col_info_tmp
         return table_columns
-        
-    def updateTableColumns(self):
+    
+    # update table_columns_tmp
+    def updateTableColumnsTmp(self):
+        table_columns_tmp = self.convertTableToTableColumns()
+        self.table_columns_tmp.setColumns(table_columns_tmp)
+    
+    # update original table_columns
+    def updateTableColumnsAndAccept(self):
         table_columns = self.convertTableToTableColumns()
         self.table_columns.setColumns(table_columns)
         self.accept()
 
     def bindFuncAllWidget(self):
-        self.widget_manager.dict_button["update"].clicked.connect(self.updateTableColumns)
+        self.widget_manager.dict_button["update"].clicked.connect(self.updateTableColumnsAndAccept)
         self.widget_manager.dict_button["exit"].clicked.connect(self.reject)
         self.widget_manager.dict_button["del_col"].clicked.connect(self.deleteSelectedTableColumns)
         self.widget_manager.dict_button["add_col"].clicked.connect(self.addNewTableColumn)
